% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.mirage_vs_result}
\alias{as.data.frame.mirage_vs_result}
\title{Coerce mirage_vs_result to data frame}
\usage{
\method{as.data.frame}{mirage_vs_result}(x, ...)
}
\arguments{
\item{x}{A \code{mirage_vs_result} object.}

\item{...}{Ignored.}
}
\value{
A data frame of variant-level results.
}
\description{
Coerce mirage_vs_result to data frame
}
