% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{log_sum_exp_vec}
\alias{log_sum_exp_vec}
\title{Log-Sum-Exp for a Vector}
\usage{
log_sum_exp_vec(v)
}
\arguments{
\item{v}{Numeric vector.}
}
\value{
log(sum(exp(v)))
}
\description{
Log-Sum-Exp for a Vector
}
\keyword{internal}
