% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirageV2-package.R
\docType{package}
\name{mirageV2-package}
\alias{mirageV2}
\alias{mirageV2-package}
\title{mirageV2: Mixture Model Based Rare Variant Analysis on Genes (v2)}
\description{
A refactored successor to the 'mirage' package.
}
\author{
\strong{Maintainer}: Anson Li \email{liyuanrui618@gmail.com}

Authors:
\itemize{
  \item Anson Li \email{liyuanrui618@gmail.com}
  \item Shengtong Han
}

}
\keyword{internal}
