% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_mirage_result}
\alias{new_mirage_result}
\title{Construct a mirage_result object}
\usage{
new_mirage_result(
  delta_est,
  eta_est,
  bf_pp_gene,
  bf_all,
  original_categories,
  call
)
}
\arguments{
\item{delta_est}{Data frame of delta estimates.}

\item{eta_est}{Data frame of eta estimates.}

\item{bf_pp_gene}{Data frame of gene-level Bayes Factors and Posterior Probabilities.}

\item{bf_all}{List of variant-level Bayes Factors per gene.}

\item{original_categories}{The original category labels before internal re-indexing.}

\item{call}{The matched call.}
}
\value{
An object of class \code{mirage_result}.
}
\description{
Construct a mirage_result object
}
\keyword{internal}
