% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_mirage_vs_result}
\alias{new_mirage_vs_result}
\title{Construct a mirage_vs_result object}
\usage{
new_mirage_vs_result(eta_est, full_info, post_prob, original_categories, call)
}
\arguments{
\item{eta_est}{Data frame of eta estimates.}

\item{full_info}{Data frame of variant-level information and Bayes Factors.}

\item{post_prob}{Data frame of variant-level Posterior Probabilities.}

\item{original_categories}{The original category labels.}

\item{call}{The matched call.}
}
\value{
An object of class \code{mirage_vs_result}.
}
\description{
Construct a mirage_vs_result object
}
\keyword{internal}
