% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mirage_vs_result}
\alias{summary.mirage_vs_result}
\title{Summary method for mirage_vs_result}
\usage{
\method{summary}{mirage_vs_result}(object, n_top = 10L, ...)
}
\arguments{
\item{object}{A \code{mirage_vs_result} object.}

\item{n_top}{Number of top variants to display. Default 10.}

\item{...}{Ignored.}
}
\description{
Summary method for mirage_vs_result
}
