% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mirage_data}
\alias{validate_mirage_data}
\title{Validate and Clean MIRAGE Input Data}
\usage{
validate_mirage_data(data, type = c("gene", "vs"))
}
\arguments{
\item{data}{A data frame containing variant counts.}

\item{type}{Either "gene" or "vs".}
}
\value{
A cleaned data frame with S3 class \code{mirage_data}.
}
\description{
Validate and Clean MIRAGE Input Data
}
\keyword{internal}
