% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mirage_params}
\alias{validate_mirage_params}
\title{Validate MIRAGE EM Algorithm Parameters}
\usage{
validate_mirage_params(
  n1,
  n2,
  gamma,
  sigma,
  eta.init,
  delta.init,
  estimate.delta,
  estimate.eta,
  fixed.eta,
  max.iter,
  tol,
  num_categories,
  type = "gene"
)
}
\description{
Validate MIRAGE EM Algorithm Parameters
}
\keyword{internal}
