% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{log_mixture_bf}
\alias{log_mixture_bf}
\title{Stable Log of (1 - eta) + eta * exp(log_bf)}
\usage{
log_mixture_bf(log_bf, eta)
}
\arguments{
\item{log_bf}{Numeric vector of log Bayes factors.}

\item{eta}{Numeric scalar or vector in \eqn{}\eqn{[0, 1]}.}
}
\value{
Numeric vector of log((1-eta) + eta * exp(log_bf)).
}
\description{
Calculates log((1-eta) + eta * BF) given log(BF), stably.
Both log_bf and eta can be vectors of the same length,
or eta can be a scalar (recycled).
}
\keyword{internal}
