% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-Sum-Exp for Two Values}
\usage{
log_sum_exp(a, b)
}
\arguments{
\item{a}{Numeric scalar.}

\item{b}{Numeric scalar.}
}
\value{
log(exp(a) + exp(b))
}
\description{
Calculates log(exp(a) + exp(b)) in a numerically stable way.
}
\keyword{internal}
